+++ CHANGELOG.md
@@ -10,6 +10,8 @@ See [keep a changelog] for information about writing changes to this log.
 
 * [PR-18](https://github.com/itk-dev/giv-din-stemme/pull/18)
   Better exceptions and `array_rand`.
+* [PR-19](https://github.com/itk-dev/giv-din-stemme/pull/19)
+  Only get published texts.
 * [PR-16](https://github.com/itk-dev/giv-din-stemme/pull/16)
   Development OIDC fixes
 * [PR-13](https://github.com/itk-dev/giv-din-stemme/pull/13)